refactor: 컨테이너 컴포넌트에 시맨틱 태그 추가 및 화면 전환 애니메이션 추가
