+++ src/components/App.jsx
@@ -1,27 +1,17 @@
-import { Outlet, useLocation } from "react-router-dom";
+import { Outlet } from "react-router-dom";
 import Nav from "./Nav/Nav";
 import Container from "./Common/Container/Container";
 import Footer from "./Footer/Footer";
 import { QueryClient, QueryClientProvider } from "@tanstack/react-query";
-import { useEffect, useState } from "react";
 
 const queryClient = new QueryClient();
 
 export default function App() {
-  const location = useLocation();
-  const [fadeKey, setFadeKey] = useState(0);
-
-  useEffect(() => {
-    setFadeKey((prev) => prev + 1);
-  }, [location.pathname]);
-
   return (
     <QueryClientProvider client={queryClient}>
       <Nav />
       <Container>
-        <div key={fadeKey} className="fade-transition">
-          <Outlet />
-        </div>
+        <Outlet />
       </Container>
       <Footer />
     </QueryClientProvider>
