+++ src/components/Common/Container/Container.jsx
@@ -1,5 +1,18 @@
-import styles from './Container.module.css';
+import styles from "./Container.module.css";
+import { useLocation } from "react-router-dom";
+import { useEffect, useState } from "react";
 
 export default function Container({ children }) {
-  return <div className={styles.container}>{children}</div>;
+  const location = useLocation();
+  const [fadeKey, setFadeKey] = useState(0);
+
+  useEffect(() => {
+    setFadeKey((prev) => prev + 1);
+  }, [location.pathname]);
+
+  return (
+    <main key={fadeKey} className={styles.container}>
+      {children}
+    </main>
+  );
 }
