+++ src/components/Common/Container/Container.module.css
@@ -1,8 +1,24 @@
 .container {
   max-width: 120rem;
   min-width: 37.5rem;
+
   width: 100%;
+
   margin: 4rem auto;
   padding-left: 2.4rem;
   padding-right: 2.4rem;
+
+  animation: fadeIn 0.7s ease;
+}
+
+@keyframes fadeIn {
+  from {
+    opacity: 0;
+    transform: translateY(10px);
+  }
+
+  to {
+    opacity: 1;
+    transform: translateY(0);
+  }
 }
