+++ src/styles/global.css
@@ -45,19 +45,3 @@ label {
   font-weight: 400;
   line-height: 1.671rem;
 }
-
-.fade-transition {
-  animation: fadeIn 0.7s ease;
-}
-
-@keyframes fadeIn {
-  from {
-    opacity: 0;
-    transform: translateY(10px);
-  }
-
-  to {
-    opacity: 1;
-    transform: translateY(0);
-  }
-}
